fix(kuma-cp) create method 'GetSplitOrdered'
