--- test/e2e/trafficroute/kuma_standalone_universal_test.go
@@ -1,1 +1,0 @@
-package trafficroute_test
